Moved GetAnnotationsDeployment() of oc and kubectl to a new function
