+++ tests/helper/helper_kubectl.go
@@ -8,8 +8,6 @@ import (
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 	"github.com/onsi/gomega/gexec"
-	applabels "github.com/openshift/odo/pkg/application/labels"
-	"github.com/openshift/odo/pkg/component/labels"
 )
 
 const (
@@ -241,21 +239,5 @@ func (kubectl KubectlRunner) VerifyResourceToBeDeleted(ri ResourceInfo) {
 // GetAnnotationsDeployment gets the annotations from the deployment
 // belonging to the given component, app and project
 func (kubectl KubectlRunner) GetAnnotationsDeployment(componentName, appName, projectName string) map[string]string {
-	var mapOutput = make(map[string]string)
-
-	selector := fmt.Sprintf("--selector=%s=%s,%s=%s", labels.ComponentLabel, componentName, applabels.ApplicationLabel, appName)
-	output := CmdShouldPass(kubectl.path, "get", "deployment", selector, "--namespace", projectName,
-		"-o", "go-template='{{ range $k, $v := (index .items 0).metadata.annotations}}{{$k}}:{{$v}}{{\"\\n\"}}{{end}}'")
-
-	for _, line := range strings.Split(output, "\n") {
-		line = strings.TrimPrefix(line, "'")
-		splits := strings.Split(line, ":")
-		if len(splits) < 2 {
-			continue
-		}
-		name := splits[0]
-		value := strings.Join(splits[1:], ":")
-		mapOutput[name] = value
-	}
-	return mapOutput
+	return GetAnnotationsDeployment(kubectl.path, componentName, appName, projectName)
 }
